requirements.txt
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
praw>=7.7.0
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0
python-telegram-bot>=20.0
slack-sdk>=3.21.0
gspread>=5.12.0
google-auth>=2.23.0
airtable-python-wrapper>=0.15.0
pydantic>=2.5.0
click>=8.1.0
colorama>=0.4.6
rich>=13.6.0
apscheduler>=3.10.0
tenacity>=8.2.0
